+++ src/Enterspeed.Source.UmbracoCms.V8/EventHandlers/EnterspeedContentCacheRefresherEventHandler.cs
@@ -28,9 +28,7 @@ namespace Enterspeed.Source.UmbracoCms.V8.EventHandlers
             IEnterspeedConfigurationService configurationService,
             IScopeProvider scopeProvider,
             IEnterspeedJobFactory enterspeedJobFactory)
-            : base(
-                umbracoContextFactory, enterspeedJobRepository, jobsHandlingService, configurationService,
-                scopeProvider)
+            : base(umbracoContextFactory, enterspeedJobRepository, jobsHandlingService, configurationService, scopeProvider)
         {
             _enterspeedJobFactory = enterspeedJobFactory;
         }
@@ -42,8 +40,8 @@ namespace Enterspeed.Source.UmbracoCms.V8.EventHandlers
 
         public void ContentCacheUpdated(ContentCacheRefresher sender, CacheRefresherEventArgs e)
         {
-            var isPublishConfigured = _configurationService.IsPublishConfigured();
-            var isPreviewConfigured = _configurationService.IsPreviewConfigured();
+            var isPublishConfigured = ConfigurationService.IsPublishConfigured();
+            var isPreviewConfigured = ConfigurationService.IsPreviewConfigured();
 
             if (!isPublishConfigured && !isPreviewConfigured)
             {
@@ -59,7 +57,7 @@ namespace Enterspeed.Source.UmbracoCms.V8.EventHandlers
 
             var jobs = new List<EnterspeedJob>();
 
-            using (var context = _umbracoContextFactory.EnsureUmbracoContext())
+            using (var context = UmbracoContextFactory.EnsureUmbracoContext())
             {
                 var umb = context.UmbracoContext;
                 foreach (var payload in jsonPayloads)
@@ -167,7 +165,6 @@ namespace Enterspeed.Source.UmbracoCms.V8.EventHandlers
 
             EnqueueJobs(jobs);
         }
-
 
         public void Terminate()
         {