+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0 - 2022-12-13]
+### Added
+  - Added virtual method to extend property data (Umbraco 8, 9, 10)
+  - Added media domain to other media types, not just images (Umbraco 7, 8, 9 & 10)
+
 ## [3.1.0 - 2022-11-22]
 ### Added
   - Added support for simple types (string, number and boolean) in grid editor array (Umbraco 7, 8, 9 & 10) (contribution by [Adrian Ochmann](https://github.com/aochmann))